requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
trafilatura==1.12.2
requests==2.31.0
lxml==5.3.0
